Backport #2037: @apollo/subgraph: remove dependency on apollo-server-types
(Note: as part of backporting, I put out another version
@apollo/cache-control-types@1.0.2 with
https://github.com/apollographql/apollo-utils/pull/173 and will now
forward-port the dependency update.)

apollo-server-types is an Apollo Server 3 package that is going away in
Apollo Server 4, so we don't want `@apollo/subgraph` to depend on it.
This was a build-time-only dependency (it just imported a type and that
type isn't used in the generated .d.ts files) but it would still be
helpful to minimize dependencies so that AS4 users don't also end up
with pieces of AS3 installed.

The new package `@apollo/cache-control-types` also has a function that
implements the "does it look like it's the right cacheControl object?"
itself, and it does not rely on AS3's `declare module` to globally
monkeypatch `cacheControl` onto `GraphQLResolveInfo`'s type definition.

Part of https://github.com/apollographql/apollo-server/issues/6057