fix @composeDirective corner case
@composeDirective wasn't working when @link of the link directive was not present.

The reason this is happening is that although we are waiting to process core schemas for think that aren't the `link` schema, we aren't doing the same for other directives. This fixes this with a slightly hacky approach of catching all exceptions that occur, and then replaying them once `completeSubgraphSchema` has been called, letting them still throw if they threw originally.