Ensures key fields needed post-@require are fetched (#2075)
Sometimes, a `@require` on some entity `E` in some subgraph A requires
that the query plan jumps from A to some other subgraph(s) to get the
requirements, and then that we use a key of `E` to jump back to A to
resume fetching the field having the `@require`. When that happens
the code was correctly finding the key need to do that "post-@require"
resuming, and was using it as "inputs" of the post-@require fetch,
but was assuming that those key fields had already fetched. And while
that's true in a number of situation, this is not guaranteed.

This fix ensures that we force fetching (pre-@require) those key
fields that are needed to resume post-@require.

Fixes #2069