gateway: update to version of gateway interface package that explicitly supports AS2
We don't really support AS2 in gateway v2 (because we require
graphql-js@16) but we want to stay up to date on this package (whose
change is more relevant for gateway v0). Also reintroduce probing for
the actual method we use on overallCachePolicy (like the comment says)
so that we actually could support AS2 a bit better if the peer dep was
ignored. (By "support" I mean "not crash", not "actually have the cache
policy feature work".)