Fix issue when type is only reachable through a @provides (#2083)
It's possible for a type within a subgraph to be only "reachable" through
a provides. Consider:
```graphql
type Query {
  t: T @provides(fields: "a { ... on A { x }}")
}

type T @key(fields: "id") {
  id: ID!
  a: A @external
}

type A {
  x: Int
}
```
In that case, type `A` is only mentioned by external field `T.a`, and
as he code was completely ignoring external fields, that type `A` was
not added to the underlying "query graph". But that type _is_ reachable
through the `@provides` on `Query.t`, and as we were trying to handle
that `@provides`, we were running into an assertion error due to `A`
not existing in the "query graph".

This patch ensures that when we have an external field, even if we don't
add any edges, we at least add the target type to the "query graph",
ensuring it exists later as we handle `@provides`.