Don't cache schema full AST by default
This was retaining a potentially non-trivial amount of memory (on large
schema, that AST is not tiny) and afaict, none of our usage was relying
on the caching much. The patch nonetheless introduce an option that
allow to easily re-enable said caching so it's easy to enabled for
specific use case later if we want to.