Fix fragment expansion nuking @defer and other minor @defer updates (#2093)
An optimisation recently made to the handling of fragments was mistakenly forgetting to look for directive applications on fragments and as a result those application were lost when that optimisation kicked in. In particular, this could make some `@defer` on a named fragments be completely ignored. This patch fixes this by simply avoiding the optimisation in question when there is applied directives, thus preserving them.

Additionally, this commit addresses comments from @glasser on #1958, namely:
- it removes/update some outdated TODOs.
- it fixes up the definition of @defer/@stream to match current spec.
- makes serialized query plan for condition hopefully a tad easier to parse (solely impact unit tests readability).
- adds a test to ensure that if the same field is queries both "normally" and deferred, we do query it twice (that's what spec specifies).