Fix issue where @defer condition gets ignored
Fixes a blunder that got overlooked: the `DeferNormalizer.init` method
had meant to be updated when condition were added to record if we had
any defer with either "no explicit label" or "some condition", and the name
of the variable keeping that information was updated, but somehow forgot
to actually set the variable for "some condition".

This wasn't caught by the existing tests because none of the condition
unit tests was using an explicit label, so the variable mentioned above
ended up being set for the "wrong" reason.

Anyway, this commit fixes the issue and adds test coverage.

Fixes #2100