Fix method removing empty branches from selection sets
As the code for @defer build the selection set used for "primary parts",
it may start adding "branches" (selections) that end up being empty
because everything within that branch is actually deferred (and thus
not part of this primary part). So once we've finished building
everything, we call a `withoutEmptyBranches` method on the selection set
that can have such (invalid for graphQL) empty branches to remove
them.

Unfortunately, the logic for that method was broken: it was removing
empty branches in simple cases, but the recursion was done the wrong
way so that if removing an empty "leaf" left its parent empty, this
wasn't removed properly. In other words, given
  `{ a b { c {} } }`
the method was returning:
  `{ a b { } }`
instead of the proper:
  `{ a }`

This commit fixes the logic of that method to behave correctly.

Fixes #2123