Fix potential assertions during query planning (#2133)
When building query plans (fetch group dependency graphs to be precise),
the method adding a group as parent of another was asserting the parent
relationship wasn't pre-existing. But in some cases, due to some
optimisations around how we handle multiple choices in planning (that
is, instead of rebuilding every possible choices from scratch, we build
everything common to every choice and then only rebuild/evaluate the part
that differs), we ended up adding parentship between groups more than
once and this was triggering the assertion mentioned above.

This commits remove the aforementioned assertion and instead make the
`addParent` method a no-op if the parent relationship already exists.