Add test for QP optimization of __typename
The optimization of #2137 is meant to have no impact on the
generated plans: it's an optimization that ensure the query
planner does not needlessly evaluate ineffient options just
because __typename happens to be resolvable from any subgraph.
But this make testing that optimization a bit more complicated.
To work around this, this patch adds a new behaviour to the
query planner whereby along the generation of the plan, it
also exposes some statistics on the plan generation. As of
this commit, the only thing exposed is the number of plan
that were evaluated under the hood, as that is what we care
to check here (but it is also one of the main contributor
to time spent query planning, so arguably an important
statistic).

There was probably more hacky ways to test this optimization,
but it while this statistic is currently only use for testing,
this feels like something that could have other uses and be
useful to enrich later.