Fix bug in patch for #2137 (#2140)
The code from #2137 is optimizing some use of __typename
when other fields are queried alongside it. But when
__typename had only inline fragments alongside it,
the logic was flawed and was discarding any potential
instance of the optimization done somewhere more
nested. This fixes that problem.

Additionally, this patch adds a test for the optimization 
of #2137, but to do that, this patch adds a new behaviour
to the query planner whereby along the generation of the
plan, it also exposes some statistics on the plan generation. 
As of this commit, the only thing exposed is the number
of plan that were evaluated under the hood, as that is 
what we care to check here (but it is also one of the main 
contributor to time spent query planning, so arguably an
important statistic).