Never merge fragments that have `@defer`
For efficiency sake, the code merges inline fragments that are "the
same" in queries. So something like:
```graphql
{
  t {
    ... on X {
      a
    }
    ... on X {
      a
      b
    }
  }
}
```
is simplified into:
```graphql
{
  t {
    ... on X {
      a
      b
    }
  }
}
```
Which is actually somewhat important because 1) while such query (the
first one) is rarely written directly, it can easily happen when you
start expanding named fragments, and 2) not doing this could create
serious inefficiencies in more complex examples as it could create
unecessary fetches.

Anyway, when directives gets involved on fragments, the code still
merge fragments but only if the directive applications are the same,
which means same name and arguments. So:
```graphql
query foo($if: Boolean) {
  t {
    ... on X @include(if: $if) {
      a
    }
    ... on X @include(if: $if) {
      a
      b
    }
  }
}
```
is simplified into:
```graphql
query foo($if: Boolean) {
  t {
    ... on X @include(if: $if) {
      a
      b
    }
  }
}
```
but:
```graphql
query foo($if1: Boolean, $if2: Boolean) {
  t {
    ... on X @include(if: $if1) {
      a
    }
    ... on X @include(if: $if2) {
      a
      b
    }
  }
}
```
is left as-is.

But those rules meant that:
```graphql
{
  t {
    ... on X @defer {
      a
    }
    ... on X @defer {
      a
      b
    }
  }
}
```
_was_ simplified by merging the 2 branches.

Except that this is kind of wrong for `@defer` as each individual is
expected to be it's own deferred chunk.

This patch special cases `@defer` so we don't such merging for it.