Fix building subgraph selections using the wrong underlying schema
In some case, we were building the selections for a subgraph based
on the "operation path", which meant the selections were incorrectly
linked to the supergraph API schema. This could lead to case where,
as we tried to add a field to such selection, the field was not
found due to being `@inacessible` and thus not part of the type
of the API schema.

This commit ensures we instead always "add" the "operation path" to
a selection set properly based on the subgraph schema we're building
the selection for.

Fixes #2152