Improves error message for misspelled override (#2181)
When mispelling the `from` of an `@override`, one can easily run into an
error about the field needing to be "shareable", which doesn't help
locate the actual problem easily. This patch detect this situation and
provide an additional hint in the error message to hopefully make
the root cause easier to locate.