Fix validation of variable on input field not taking default into account (#2176)
Some typo made it so that when validating usages of a variable in a query, if the variable was used for the value of an input field, the default of that input field was not involved in the validation. And so a variable with a nullable type used on a field of non-nullable type would be rejected even when there is actually a default value.