Optimize plan for defer where only keys are fetched (#2182)
Router-based defer is based on using an entity key to "resume" the
deferred part of a `@defer`, but that means that at least the key(s)
needs to be fetched in the non-deferred part. But if only the key
were selected in the deferred part, the code was issuing a new
deferred fetch to get those key field "again". This commit avoids
this by recognizing that those key fields don't need to be fetched
again.