Fix building additional abstract type edges
As we build the query graph, some edges are added between abstract types
to avoid type explosion when we can. Unfortunately, in some cases,
adding those edge was incorrect and we need to type-explode due to
some specific different between subgraphs on the definition of the
types involved. This fixes this issue.

Fixes #361