Fix error when a skipped enum value had directives applied
We sometimes have to call the method merging directives on a
merged value that we end up removing later, and this because
we need @inaccessible to be merged to decide if the value
should be kept (this is case of enum value, but I think input
field are also the same).

But since #2136, most directives (@inaccessible being the one
exception) are not merged right away, but we simply "record"
that they should be merged and merge them later. The result
however is cases where when we later try the merging, the
element is not attached to the schema anymore, and an assertion
was triggered. This fixes this cases.