Fix potential QP issue with shareable root fields (#2239)
On some cases, during query planning, the query planner had to consider plans whose dependencies where essentially:
```
`Root Fetch 1`
                         \
                          `Dependent fetch`
                         /
`Root Fetch 2`
```
but the "processing" code wasn't expecting that and that triggered an assert (the patch has a unit test that, if run against pre-patch `main` triggers that assertion; note that the assertion is thrown during query planning, but the final query plan of that test does _not_ look like the one above). The patch fixes this issue.