Loosen validation of @shareable on interface fields
The `@shareable` directive was always intended to be only on object
types. However, its location being `FIELD_DEFINITION`, it is
syntactically allowed on interface fields, and due to an oversight of
the original code, doing so had not been propertly rejected (but it
had no effects whatsoever).

This oversight was noticed during #2175 and a validation was added
to reject a `@shareable` on interface fields. Unfortunately, it appears
a number of users had started putting such `@shareable` instances,
so the added validation breaks composition for those users on upgrades
(the "fix" for those user is trivial, just removing `@shareable` on
any interface field, it did nothing and so will have no other impact).

However, sticking with the status quo is highly undesirable: continuing
to allow `@shareable` on interface fields but having it do nothing
whatsoever is guaranteed to create confusion, and can even be harmful
if a user is falsly led to believe that it does something it does not.

So this patch explores a middle-ground alternative: it gives a meaning
to `@shareable` on interface fields. That meaning is to force
implementations to also mark the field as `@shareable`, as this feel
like the most intuitive behaviour that can be done simply. Note that
this is a middle ground in the sense that a subset of the users that had
started using `@shareable` on interface fields may still see composition
break on upgrades, if their implementation does not "manually inherit"
those `@shareable` already. But experimentally, it appears that this
patch avoids breaking for almost all the cases we have access to.