Don't validate out introspection type names
The fix for #2237 has been a bit over-zealous, at least to some extent,
as `graphql-js` accepts a schema that manually redeclare the introspection
types but those are now rejected.

It's a bit unclear why `graphql-js` makes that particular exception, in
particular since:
1. it makes no such special cases for the introspection fields (being it
   `__typename`, `__schema` or `type(name)`). Particuarly strange when
   `__schema` and `__type` are the actual "entry point" for those
   introspection types in the first place ...
2. it does not seem to otherwise validate that those introspection types
  are valid, though "manual" definitions seems to be silently discarded
  so have no impact.

But in any case, being consistent with `graphql-js` makes sense, so this
patch re-allow definitions of introspection types, but ignores them.