Fix query planning not find plan in some specific interactions of shareable fields and interfaces (#2246)
When the field of an interface is queried, the query planner has 2
options it can theory consider:
1. it can just fetch the interface field directly.
2. or it can type-explode the interface into all of its implementation
  and try to fetch the field on each of them.

Of course the first option is often best, but there is some special
corner cases where that's not true, specifally:
a. when the field has a @provides on some implementations. When that is
  the case, it is possible that type-exploding and making use of that
  @provides yield a most efficient query plan. In other words, both
  options above work, but the 2nd one turns out to be more efficient.
b. if a field `s` of an interface `I` leads to some some shared value
  type `S` and that interface is implemented by an entity type `E`, then
  it's possible that a field of `S` is not reachable at all if we
  take `f` from `I` directly (it's not in the "current" subgraph),
  but type-exploding allows to use a @key on `E` to another subgraph
  and allow to reach that `S` field. In other words, it's a case
  where option 1 above does not end up working, but option 2 does.

Case a. above was correctly handled by the current code, but case b. was
not. Instead, the code was not trying to type-explode, and thus was not
finding a plan when one actually exists (and composition validation did
succeed).

Note that the slight difficulty here is that while it would be
simple to have the query planner _always_ try type-exploding (so
always check both option 1 and 2 above), doing so create more options
to consider which leads to slower query planning time (and possibly
non-optimal planning in some cases). So the patch introduces a
a few mechanisms to recognize some subset of case 2.b that are
not problematic (and avoid type explosion for those) _and_ ensure
that if we do type explode, we still at least save time later
when computing actual query plans by ignoring the type-explosion
case if the "direct" option ended up successful.