Avoid reusing named fragments that are invalid for the subgraph (#2255)
In some subtle conditions (mostly related to interface merging), it is
possible that a named fragment from the user query "matches" in a
subgraph fetch, yet to have that fragment being invalid for the
subgraph (see test included in this patch for an example). This
patch ensures we don't reuse such fragments.