Fix issue with path in query plan's deferred nodes
The `path` currently included in the query plan for deferred nodes is a
"response" path, meaning that it's essentially a path in a json object.
But for deferred node, the idea was that the path, alongside the
`subselection` that the deferred node also include, could be used to
rebuild the subset of the query corresponding to the deferred part it
represents, but because `path` misses fragments, this is not always
possible to do.

This patch changes the `path` format to include fragments so it can be
used to rebuild queries more directly (the new format also do not
include `@` for lists: this gets in the way of rebuilding queries from
the path more than it helps, and ever when using the `path` to walk
on responses, it's easy enough to handle lists without needing this).