Adds test that validates that older supergraphs can be read by new versions
Because newer versions can't generate old subgraphs, this commit creates
a new scripts that allow to easily generate a supergraph for the
"current" version. The intent is that with each new version, we'll
generate the corresponding supergraph and commit it to source control.
As such, any following version will have older versions of the
supergraph which it can test the reading of.

This is imperfect as this only test a relatively specific supergraph,
and making change to it would be painful since we've have to backport
the change to older versions and re-generate new versions of the
supergraph (possible, but painful). But it is better than nothing as
it will quickly catch any obvious breaking in the reading of older
supergraphs.