Generates condition nodes for @include/@skip when possible
When a fetch to a subgraph is entirely included inside some `@include`/`@skip`,
we can use the `ConditionNode` introduced for `@defer` to completely
save the fetch when it's not included/is skipped.

This commit detects when a fetch can be "wrapped" in one or more
`ConditionNode` due to `@include`/`@skip` and generates the plan
accordingly.

Additionally, this commit adds support for `ConditionNode` in the
gateway execution code, since it was not handled before.

Fixes #2251