Fixes composition issues with @interfaceObject (#2318)
This commit fixes two main problems:
1. merging a shareable field whose return type was an `@interfaceObject`
   was not working properly.
2. some metadata was not properly computed on a path added for
   `@interfaceObject` and this led to a potential assertion errors
   during composition validation on some cases.

For the merging of shareable fields, the main issue was that the code
checking if 2 types are "the same" was relying on the type "kind", which
was fine before but broke with `@interfaceObject` in the sense that some
type type T can be an interface type in one subgraph but an object type
in another (through `@interfaceObject`) and we were not considering
those "the same type". But with that fixed, a small issue remained in
that the check added in #1556 had not been updated for `@interfaceObject`
and needed to ignore those.

For the 2nd issue, the core bug was that when building a `GraphPath`, we
keep track of the edge that made us enter the current graph, and some
assertion is essentially checking that there is no "key" edges between
that edge and the end of the path, but that "edge entering the subgraph"
info was not properly set on a path added for `@interfaceObject` and
that triggered the assertion. This commit fix this issue, but this was
also highlighting some ineffeciency whereby we were generating
unecessary paths (path moving to a subgraph and then moving back again
right away), so the patch also fix that inefficiency.