Improve generation of plans once all path options are computed (#2316)
At a very high level, query planning works in 2 main steps:
1. for every "leaf" of the query, we compute which possible options
   (paths) can lead to it.
2. we take all those options and generates the corresponding plans,
  which means we compute all the (non trivially inefficient) possible plan
  that would be valid for the query. For each of those plan, we compute
  the cost of the plan and keep the best plan.

In some cases, the 1st step gives quite a few possible options, and
generating all the corresponding plans is costly.

This patch attempts to first generate plans that are a bit more likely
to be the most efficient, or at least fairly good. When then use
the cost of those plans to something cut the generation of other plans
early. Essentially, as we generate a plan corresponding to a set of
options, if we notice that the cost gets higher than the best we've
found so far before we've handled all the paths, then we can give up
on that set of options without generating everything.

Doing so have little to no impact on very simple plans, but for more
complex ones it can drastically reduce the generation time
(anecdotically, some real world plan gets more than a 10x improvement on
generation speed).