Rewrites gateway post-processing
With `@interfaceObject`, some queries can sometimes be entirely
fullfilled by a sugraph using the `@interfaceObject` (when only
fields provided by that subgraph are queried and the query does
not necessitate knowledge of the underlying implementation type
as neither `__typename` is queried, nor is there any type
condition). But the post-processing that the gateway does on
the internal results to compute the final response was currently
using GraphQL-js execution, and this was breaking in those cases
because we get results for what is an interface in the supergraph
but with no __typename.

Additionally, using GraphQL-js execution for this post-processing
has long been a "FIXME" of the code, as this is a heavy-handed way
to do this processing.

This patch rewrite this post-processing to stop using GraphQL-js
execution and instead implements the subset of what is truly
necessary for that post-processing. Doing so allows to ensure that
the code work as needed with the `@interfaceObject` cases mentioned
above.

Fixes #2320